refactor: SocialNetwork 상세 화면 패키지 분리 및 관련 로직 개선
- `SocialNetworkDetailScreen` 및 `SocialNetworkDetailContent`를 `socialnetwork` 하위 패키지로 이동
- `AfternoteDetailSuccessMapper`에서 `ReceiverUiModel`의 ID가 없을 경우 인덱스 기반의 기본값(`receiver_$index`)을 할당하도록 수정
- `AfternoteDetailSuccessMapper` 내 `toDetailContentUiModel` 함수에 가독성을 위한 중괄호 블록 추가
- 패키지 구조 변경에 따라 `AfternoteNavGraphDetail` 및 관련 스크린(`Gallery`, `Memorial`)의 임포트 경로와 Kdoc 참조 업데이트
- `SocialNetworkDetailScreen`에 필요한 상세 상태 관련 클래스들의 임포트 추가 및 정리
