+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/navigation/AfternoteNavGraphDetail.kt
@@ -19,7 +19,7 @@ import com.afternote.feature.afternote.presentation.R
 import com.afternote.feature.afternote.presentation.author.detail.AfternoteDeleteState
 import com.afternote.feature.afternote.presentation.author.detail.GalleryDetailRoute
 import com.afternote.feature.afternote.presentation.author.detail.MemorialGuidelineDetailRoute
-import com.afternote.feature.afternote.presentation.author.detail.SocialNetworkDetailRoute
+import com.afternote.feature.afternote.presentation.author.detail.socialnetwork.SocialNetworkDetailRoute
 import com.afternote.feature.afternote.presentation.author.navigation.model.AfternoteRoute
 
 @Composable
